+++ internal/cli/access_policy.go
@@ -173,6 +173,7 @@ jira issue comment list
 jira issue create
 jira issue delete
 jira issue edit
+jira issue field get
 jira issue field set
 jira issue fields
 jira issue get