+++ internal/cli/output_contract.go
@@ -83,6 +83,7 @@ jira issue children
 jira issue comment list
 jira issue create
 jira issue edit
+jira issue field get
 jira issue field set
 jira issue fields
 jira issue get